# rendu/clement/variance_stddiv/openapi.yaml
openapi: 3.1.0
info:
  title: Clement Enterprise Population Statistics API
  version: 1.0.0
  description: API HTTP optionnelle autour des fonctions 3, 7 et 9 du cahier des charges.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Verification de sante
      responses:
        '200':
          description: Service disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /variance:
    post:
      summary: Calcule la variance de population
      operationId: variancePopulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatisticsRequest'
      responses:
        '200':
          description: Resultat du calcul
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
        '422':
          description: Les nombres sont invalides ou le résultat dépasse la plage numérique
  /ecart-type:
    post:
      summary: Calcule l'ecart type de population
      operationId: ecartTypePopulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatisticsRequest'
      responses:
        '200':
          description: Resultat du calcul
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
        '422':
          description: Les nombres sont invalides ou le résultat dépasse la plage numérique
  /moyenne:
    post:
      summary: Calcule la moyenne
      operationId: moyennePopulation
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [standard, ilian]
            default: standard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatisticsRequest'
      responses:
        '200':
          description: Resultat du calcul
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
        '422':
          description: Les nombres sont invalides ou le résultat dépasse la plage numérique
        '503':
          description: Le fichier Ilian est absent ou invalide pour le mode ilian
components:
  schemas:
    StatisticsRequest:
      type: object
      required: [nombres]
      properties:
        nombres:
          type: array
          items:
            type: number
            format: double
          example: [12, 4.5, 3]
    StatisticsResponse:
      type: object
      required: [operation, resultat, cached]
      properties:
        operation:
          type: string
          enum: [moyenne_population, variance_population, ecart_type]
        resultat:
          type: [number, 'null']
          format: double
        cached:
          type: boolean
        mode:
          type: [string, 'null']
          enum: [standard, ilian, null]